\name{colMT}
\docType{data}
\alias{colMT}
\title{MT map color scheme}
\description{
   Color map implementing the color scheme for MT maps.
    This is the plasma scale from Matplotlib (pyplot)
    generated by function plasma from package \pkg{viridisLite}.
}
\usage{colMT}
\format{A vector with 256 RGB color values.}
\keyword{datasets}
