\name{qMRI-package}
\alias{qMRI-package}
\alias{qMRI}
\docType{package}
\title{
\packageTitle{qMRI}
}
\description{
\packageDescription{qMRI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{qMRI}
\packageIndices{qMRI}
}
\author{
Karsten Tabelow \email{tabelow@wias-berlin.de}\cr
J\"org Polzehl \email{polzehl@wias-berlin.de}

Maintainer: \packageMaintainer{qMRI}
}
\references{
Weiskopf, N.; Suckling, J.; Williams, G.; Correia, M. M.; Inkster, B.; Tait, R.;
Ooi, C.; Bullmore, E. T. & Lutti, A. Quantitative multi-parameter mapping of
R1, PD(*), MT, and R2(*) at 3T: a multi-center validation. Front Neurosci,
Wellcome Trust Centre for Neuroimaging, UCL Institute of Neurology, University College
London, UK., 2013, 7, 95
}
\keyword{ package }
\seealso{
\code{\link[aws:aws-package]{aws}}
}
