\name{qMRI-package}
\alias{qMRI-package}
\alias{qMRI}
\docType{package}
\title{
Analysis methods for quantitative Magnetic Resonance Imaging
}
\description{
Provides functions to handle and analyze data from quantitative Magnetic Resonance Imaging experiments.
Special reference is given to Multi-Parameter Mapping (MPM)  
}
\details{
\tabular{ll}{
Package: \tab qMRI\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2017-04-02\cr
License: \tab GPL (>= 2)\cr
}
Eperimental data can be read using function \code{\link{readMPMData}}. Evaluation of the 
ESTATICS model (Weisskopf (2013) is provided by functions \code{\link{estimateESTATICS}}
and \code{\link{estimateESTATICSQL}} with the first implementing nonlinear least squares regression 
and the second a quasi-likelihood approach assuming a noncentral chi- or a Rician distribuion
for the data. The latter function should be preferred in case of low SNR (high resolution) 
data to avoid biased parameter estimates. Function \code{\link{estimateESTATICSQL}} requires
a specification of the scale parameter sigma of the data distribution. For the case of 
the Rician distribution paramter maps of sigma can be estimated from magnitude and phase images 
(complex polar coordinates) using function \code{\link{estimateSigma}}. Function \code{\link{smoothESTATICS}}
implements (vector) adaptive smoothing for parameter maps generated by \code{\link{estimateESTATICS}}
and \code{\link{estimateESTATICSQL}}. Functions \code{\link{calculateQI}} 
and \code{\link{estimateQIconf}} calculate estimates of quantitaive imaging parameters (R1, R2star, PD and 
(if a Magnetic Transfer Series was used) delta) and corresponding confidence information from the 
parameters in the ESTATICS model. Functions \code{\link{writeEstatics}}, \code{\link{writeQI}} and \code{\link{writeQIconf}} 
export the results as Nifti-Files.
}
\author{
Karsten Tabelow \email{karsten.tabelow@wias-berlin.de}, Joerg Polzehl \email{joerg.polzehl@wias-berlin.de}



Maintainer: Karsten Tabelow \email{karsten.tabelow@wias-berlin.de}
}
\references{
Weiskopf, N.; Suckling, J.; Williams, G.; Correia, M. M.; Inkster, B.; Tait, R.; Ooi, C.; Bullmore, E. T. & Lutti, A. Quantitative multi-parameter mapping of R1, PD(*), MT, and R2(*) at 3T: a multi-center validation. Front Neurosci, Wellcome Trust Centre for Neuroimaging, UCL Institute of Neurology, University College London London, UK., 2013, 7, 95
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
\code{\link{readMPMData}},\code{\link{estimateESTATICS}},\code{\link{estimateESTATICSQL}},\code{\link{estimateSigma}},\code{\link{smoothESTATICS}},
\code{\link{calculateQI}},\code{\link{estimateQIconf}},\code{\link{writeEstatics}}, \code{\link{writeQI}},\code{\link{writeQIconf}} 
}
\examples{
~~ simple examples of the most important functions ~~
}
